Moving tests in their correct place
