+++ tests/community_specific/combine_object/test_combine_archive.py
@@ -1,13 +1,14 @@
 import pytest
 
-from app.models import CombineArchive
+from app.models import CombineArchive, CombineArchiveException
 
 
 @pytest.mark.parametrize(
     "filename",
     [
         "tests/data/omex/CombineArchiveShowCase.omex",
-        "tests/data/omex/Roda2020.omex",
+        "tests/data/omex/case_01.omex",
+        "tests/data/omex/Elowitz_Leibler_2000.omex",
     ],
 )
 def test_combine_archive_init_from_archive(filename):
@@ -36,3 +37,24 @@ def test_combine_archive_init_from_sbml(filename):
     assert ca.main_model_object is not None
     assert ca.main_model_location is None
     assert ca.main_model_metadata is not None
+
+
+def test_copasi_metadata():
+    with pytest.raises(CombineArchiveException) as exc:
+        CombineArchive("tests/data/omex/Roda2020.omex")
+    assert "COPASI files are not yet handled by the FairCombine assessment tool" in str(
+        exc.value
+    )
+
+
+@pytest.mark.parametrize(
+    "filename",
+    [
+        "tests/data/omex/Ai2021PhysiomeS000013.omex",
+        "tests/data/omex/SMC_excitation_contraction.omex",
+    ],
+)
+def test_no_master_file_in_manifest(filename):
+    with pytest.raises(CombineArchiveException) as exc:
+        CombineArchive(filename)
+    assert "No master file was defined in provided archive manifest" in str(exc.value)
